requirements.txt
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
chromadb
torch
protobuf==3.20.3


